update evaluation code
